:sparkles: find_module: 上新：每日推荐图标可以点击了，页面也写好了！补充了一点其它的细节

+++ module_find/src/main/java/com/handsome/module/find/view/viewmodel/SpecialEditionViewModel.kt
@@ -15,7 +15,7 @@ class SpecialEditionViewModel : ViewModel() {
     val stateFlow : StateFlow<AlbumData?>
         get() = _mutableAlbumStateFlow.asStateFlow()
 
-    fun getAlbumData(id : Int){
+    fun getAlbumData(id : Long){
         viewModelScope.launch {
             _mutableAlbumStateFlow.emit(FindAlbumApiService.INSTANCE.getAlbumData(id))
         }
